requirements.txt
PyMuPDF
pytesseract
pdf2image
Pillow
